[update] improved api route readability

+++ server/app/Http/Controllers/AdminCorporate/AffiliationController.php
@@ -18,6 +18,8 @@ class AffiliationController extends Controller
      */
     public function index()
     {
+        Gate::authorize('check-membership', [['admin', 'corporate']]);
+
         $order = request()->input('order', 'asc');
         $per_page = request()->input('per_page', '10');
         $sort = request()->input('sort', 'id');
@@ -37,6 +39,8 @@ class AffiliationController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('check-membership', [['corporate']]);
+
         $valid = $request->validate([
             'name' => 'required|string|unique:affiliations,name',
             'priority' => 'required|numeric|min:1|unique:affiliations,priority',
@@ -58,6 +62,8 @@ class AffiliationController extends Controller
      */
     public function update(Request $request, Affiliation $affiliation)
     {
+        Gate::authorize('check-membership', [['corporate']]);
+
         $valid = $request->validate([
             'name' => 'required|string|unique:affiliations,name,' . $affiliation->id,
             'priority' => 'required|numeric|min:1|unique:affiliations,priority,' . $affiliation->id,
@@ -78,6 +84,8 @@ class AffiliationController extends Controller
      */
     public function destroy(string $affiliation)
     {
+        Gate::authorize('check-membership', [['corporate']]);
+
         $ids = explode(",", $affiliation);
         $deleted_count = \App\Models\Affiliation::destroy($ids);
