+++ server/app/Http/Controllers/AdminCorporate/MailTemplateController.php
@@ -7,16 +7,21 @@ use App\Http\Requests\MailTemplateStoreUpdateRequest;
 use App\Models\MailTemplate;
 use App\Services\MailTemplateService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class MailTemplateController extends Controller
 {
     public function index(MailTemplateService $service)
     {
+        Gate::authorize('check-membership', [['admin', 'corporate']]);
+
         return $service->list();
     }
     
     public function store(MailTemplateStoreUpdateRequest $request, MailTemplateService $service)
     {
+        Gate::authorize('check-membership', [['admin', 'corporate']]);
+
         $service->store($request);
 
         return response()->json([
@@ -26,6 +31,8 @@ class MailTemplateController extends Controller
     
     public function update(MailTemplateStoreUpdateRequest $request, MailTemplate $mail_template, MailTemplateService $service)
     {
+        Gate::authorize('check-membership', [['admin', 'corporate']]);
+
         $service->update($request, $mail_template);
 
         return response()->json([
@@ -35,6 +42,8 @@ class MailTemplateController extends Controller
     
     public function massDelete(string $ids, MailTemplateService $service)
     {
+        Gate::authorize('check-membership', [['admin', 'corporate']]);
+
         $deleted_count = $service->deleteIds($ids);
 
         return response()->json([
@@ -44,6 +53,8 @@ class MailTemplateController extends Controller
     
     public function massUpdate(Request $request, MailTemplateService $service)
     {
+        Gate::authorize('check-membership', [['admin', 'corporate']]);
+
         $count = $service->updatePriorities($request);
 
         return response()->json([
